Simplify DSL helper method: vault_parameter!
This simplifies the vault helper method provided to the template DSL to simply
insert a named parameter with a generic description and NoEcho set. This
parameter uses a custom type that is not recognized by AWS and will cause a
validation error if not replaced.

The template callback will handle finding all parameters of this custom type,
generating values for them and saving them safely in Vault, either to an
appropriate project path or to the cubbyhole for the current token. The type is
then converted to 'String' and the value is filled from Vault during stack
creation.